refactor: move directory model code
