+++ endpoint.py
@@ -1,8 +1,51 @@
 from fastapi import FastAPI, Depends, HTTPException, Query, status
-from sqlmodel import Session, select
-from models.model_and_session import *
+from sqlmodel import select, Field, SQLModel
+from session_def import *
 from typing import List
-# sqlmodel_databaseからインポートしたモデルからテーブルを生成
+import pytz
+from datetime import datetime
+from typing import Optional
+
+# sqlモデルクラス定義
+
+class TaskBase(SQLModel):
+    # baseモデル　titleのみbaseモデルでも必須値　notionのように
+    title:                   str
+    description:    Optional[str] = None
+    progress_state: Optional[int] = None
+    limit_at:  Optional[datetime] = None
+
+
+
+
+
+class Task(TaskBase, table=True):
+
+    # 実際にテーブルとして追加するモデル
+    id: Optional[int] = Field(default=None, primary_key=True)
+
+    # インスタンス作成時に自動作成される時間のデータ
+    created_at: datetime = Field(default_factory=lambda: datetime.now(pytz.timezone('Asia/Tokyo')), nullable=False)
+    updated_at: datetime = Field(default_factory=lambda: datetime.now(pytz.timezone('Asia/Tokyo')), nullable=False)
+    __tablename__ = 'tasks'
+
+
+class TaskCreate(TaskBase):
+    pass
+
+
+class TaskRead(TaskBase):
+    pass
+
+
+class TaskUpdate(SQLModel):
+
+    title:          Optional[str] = None
+    description:    Optional[str] = None
+    progress_state: Optional[int] = None
+    limit_at:       Optional[datetime] = Field(default_factory=lambda: datetime.now(pytz.timezone('Asia/Tokyo')), nullable=True)
+
+
 
 
 app = FastAPI()
